requirements_real.txt
# CrewAI Compliance Experts - Real Implementation Dependencies
# Use this file for production with actual CrewAI framework

# Core CrewAI Framework
crewai[tools]>=0.51.0

# LLM Integrations (choose based on your provider)
# For Ollama (Free, Local) - RECOMMENDED
langchain-ollama>=0.1.0
ollama>=0.3.0

# For OpenAI (API Key Required)
langchain-openai>=0.1.0  
openai>=1.0.0

# For Anthropic Claude (API Key Required)
langchain-anthropic>=0.1.0
anthropic>=0.25.0

# Core Dependencies (same as before)
fastapi>=0.111
uvicorn[standard]>=0.30
chromadb>=0.5
sentence-transformers>=3.0
duckdb>=1.0
pandas>=2.2
networkx==3.4.2
pydantic>=2.8
python-dotenv>=1.0
PyYAML>=6.0

# Additional Production Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
